# project/static/swagger.json
swagger: '2.0'
info:
  description: This is an API developed to test some CRUD operation based on multi user register and login, authenticated with JWT.
  version: 1.0.0
  title: CRUD Flask API
  # put the contact info for your development or API team
  contact:
    email: cdchinmoy@gmail.com

  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html

# tags are used for organizing operations
...
paths:
...
  /doctor_api/add_medicine:
    post:
      tags:
      - Doctor Endpoint
      summary: log doctor into the system
      operationId: Medicine
      description: Allows a legitimate doctor access privileged info
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - in: query
        name: token
        required: true
        schema:
          type: string
        description: token to verify user authentication
      - in: body
        name: medicine
        description: Doctor to be logged in
        required: true
        schema:
          $ref: '#/definitions/Medicine'
      responses:
        200:
          description: With json object that has "success" and "token" attributes signify a successful login request. With json object that has "error" attribute signifies a failed login request
...
definitions:
...
  Medicine:
    type: object
    required:
    - title
    - dose
    properties:
      title:
        type: string
        example: Paracetamol
      price:
        type: number
        example: 3
...
host: localhost:5000
schemes:
 - http